creamos la carpeta helpers y acomodamos el codigo para que sea mas entendible
